Before:

/var/www # php artisan auto:cache_random_media
41148 movie ids and 10164 tv ids cached. Movie query time: 897 ms. Movie cache time: 15 ms. /var/www # php artisan auto:cache_random_media
41148 movie ids and 10164 tv ids cached. Movie query time: 842 ms. Movie cache time: 14 ms. /var/www # php artisan auto:cache_random_media
41148 movie ids and 10164 tv ids cached. Movie query time: 842 ms. Movie cache time: 13 ms.

After:

/var/www # php artisan auto:cache_random_media
13948 movie ids and 10164 tv ids cached. Movie query time: 155 ms. Movie cache time: 8 ms. /var/www # php artisan auto:cache_random_media
13948 movie ids and 10164 tv ids cached. Movie query time: 192 ms. Movie cache time: 6 ms.